Refactor : 투표 저장 로직 수정(#23)
